feat: add task history panel and integrate with task detail and list components
