+++ TaskFlow.Web/src/pages/TaskDetailPage.tsx
@@ -1,12 +1,12 @@
 import { useState } from 'react'
 import { useParams, Link, useNavigate } from 'react-router-dom'
-import { useTaskQuery, useTaskHistoryQuery, useUpdateTaskMutation, useDeleteTaskMutation } from '@/hooks/useTasks'
+import { useTaskQuery, useUpdateTaskMutation, useDeleteTaskMutation } from '@/hooks/useTasks'
 import { useNotesQuery, useCreateNoteMutation, useUpdateNoteMutation, useDeleteNoteMutation } from '@/hooks/useNotes'
+import { TaskHistoryPanel } from '@/components/TaskHistoryPanel'
 import { TaskForm } from '@/components/TaskForm'
 import { NoteCard } from '@/components/NoteCard'
 import { NoteForm } from '@/components/NoteForm'
 import { formatDate } from '@/lib/utils'
-import type { TaskItemEventResponseDto } from '@/api/tasks'
 import { formatShort } from '@/lib/journal-utils'
 import type { TaskItemResponseDto, NoteResponseDto, CreateTaskItemDto } from '@/api/client/types.gen'
 import '@/tasks.css'
@@ -22,7 +22,6 @@ export function TaskDetailPage() {
   const taskId = Number(id)
 
   const { data: taskData, isLoading: taskLoading, error: taskError } = useTaskQuery(taskId)
-  const { data: historyData, isLoading: historyLoading } = useTaskHistoryQuery(taskId)
   const { data: notesData, isLoading: notesLoading } = useNotesQuery(taskId)
   const updateTask = useUpdateTaskMutation()
   const deleteTask = useDeleteTaskMutation()
@@ -38,7 +37,6 @@ export function TaskDetailPage() {
   const [taskMutationError, setTaskMutationError] = useState<string | null>(null)
 
   const task = taskData?.data as TaskDetailModel | undefined
-  const history: TaskItemEventResponseDto[] = (historyData?.data as TaskItemEventResponseDto[] | undefined) ?? []
   const notes: NoteResponseDto[] = (notesData?.data as NoteResponseDto[] | undefined) ?? []
 
   if (!Number.isFinite(taskId)) return <div className="tasks-page"><div className="t-shell"><p className="t-error">Task not found.</p></div></div>
@@ -113,23 +111,7 @@ export function TaskDetailPage() {
           <div className="t-section-hdr">
             <h2 className="t-section-title">Task history</h2>
           </div>
-
-          {historyLoading ? (
-            <p className="t-empty">Loading history…</p>
-          ) : history.length === 0 ? (
-            <p className="t-empty">No history yet.</p>
-          ) : (
-            <div className="t-panel">
-              <ol className="t-history-list">
-                {history.map(event => (
-                  <li key={event.id} className="t-history-item">
-                    <div className="t-history-line">{describeEvent(event)}</div>
-                    <div className="t-history-time">{formatDateTime(event.occurredAtUtc)}</div>
-                  </li>
-                ))}
-              </ol>
-            </div>
-          )}
+          <TaskHistoryPanel taskId={taskId} />
         </div>
 
         <div>
@@ -182,42 +164,6 @@ export function TaskDetailPage() {
       </div>
     </div>
   )
-}
-
-function describeEvent(event: TaskItemEventResponseDto): string {
-  if (event.changeSummary) return event.changeSummary
-
-  switch (event.eventType) {
-    case 'TaskCreated':
-      return 'Task was created'
-    case 'AssignedToJournalDay':
-      return event.toJournalDate ? `Assigned to ${formatShort(event.toJournalDate)}` : 'Assigned to a journal day'
-    case 'ReassignedToJournalDay':
-      return event.fromJournalDate && event.toJournalDate
-        ? `Moved from ${formatShort(event.fromJournalDate)} to ${formatShort(event.toJournalDate)}`
-        : 'Moved to a different journal day'
-    case 'RemovedFromJournalDay':
-      return event.fromJournalDate ? `Removed from ${formatShort(event.fromJournalDate)}` : 'Removed from a journal day'
-    case 'Completed':
-      return 'Marked complete'
-    case 'Reopened':
-      return 'Reopened'
-    case 'TitleChanged':
-      return 'Title updated'
-    case 'PriorityChanged':
-      return 'Priority changed'
-    case 'StatusChanged':
-      return 'Status changed'
-    default:
-      return event.eventType
-  }
-}
-
-function formatDateTime(value: string): string {
-  return new Intl.DateTimeFormat('en-US', {
-    dateStyle: 'medium',
-    timeStyle: 'short',
-  }).format(new Date(value))
 }
 
 function getTaskMutationErrorMessage(error: unknown): string {
