+++ TaskFlow.Web/src/tasks.css
@@ -59,10 +59,51 @@
   color: var(--ink);
 }
 
+.t-panel-title--tight {
+  margin-bottom: 4px;
+}
+
 .t-inline-error {
   margin: 0 0 12px;
   font-size: 13px;
   color: var(--danger-ink);
+}
+
+.t-modal-backdrop {
+  position: fixed;
+  inset: 0;
+  background: rgba(15, 23, 42, .28);
+  backdrop-filter: blur(2px);
+  z-index: 40;
+}
+
+.t-modal {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: min(720px, calc(100vw - 32px));
+  max-height: calc(100vh - 48px);
+  overflow: auto;
+  background: var(--paper);
+  border: 1px solid var(--line);
+  border-radius: var(--radius);
+  box-shadow: var(--shadow-md);
+  padding: 22px 26px;
+  z-index: 41;
+}
+
+.t-modal-header {
+  display: flex;
+  align-items: flex-start;
+  justify-content: space-between;
+  gap: 16px;
+  margin-bottom: 16px;
+}
+
+.t-modal-subtitle {
+  font-size: 13px;
+  color: var(--muted);
 }
 
 /* ─── Filter bar ──────────────────────────────────────────────────────────── */
